Backend authentication using JWT

+++ backend/src/controllers/posts.controller.js
@@ -63,6 +63,7 @@ export async function getAllPublicPosts(req, res) {
 }
 /**
  * Get all posts of a given author
+ * TODO: Get public and friend posts only
  * @param {Express.Request} req
  * @param {Express.Response} res
  * @returns All the posts of an author
@@ -124,6 +125,7 @@ export async function getAllPosts(req, res) {
 
 /**
  * Get one post of a given author (using author id)
+ * TODO: Check if user is allowed to see post (public/friend, unlisted)
  * @param {Express.Request} req
  * @param {Express.Response} res
  * @returns The post information requested
@@ -185,8 +187,14 @@ export async function getOnePost(req, res) {
  * @returns New post
  */
 export async function putPost(req, res) {
+	// Check if the user is posting to their own author page
+	if (req.user.id !== req.params.authorId) {
+		return res.status(403).json({ error: 'Cannot post on this profile' });
+	}
+	
 	const post = req.body;
 	post.id = req.params.id;
+	post.authorId = req.user.id;
 	post.published = new Date();
 
 	if (!validPost(post))
@@ -204,9 +212,17 @@ export async function putPost(req, res) {
  */
 export async function deletePost(req, res) {
 	const id = req.params.id;
+	const authorId = req.params.authorId;
+
 	if (!id) {
 		return res.status(400).json({ error: 'Missing required property' });
 	}
+
+	// Check if the user is allowed to delete the post
+	if (authorId !== req.user.id) {
+		return res.status(403).json({ error: 'Not allowed' });
+	}
+
 	await postService.deletePost(id);
 	return res.status(204);
 }
@@ -220,6 +236,12 @@ export async function deletePost(req, res) {
 export async function updatePost(req, res) {
 	const post = req.body;
 	post.id = req.params.id;
+	const authorId = req.params.authorId;
+
+	// Check if the user is allowed to update the post
+	if (authorId !== req.user.id) {
+		return res.status(403).json({ error: 'Not allowed' });
+	}
 
 	if (!validPost(post)) {
 		return res.status(400).json({ error: 'Missing required property' });
@@ -236,7 +258,13 @@ export async function updatePost(req, res) {
  * @returns New post
  */
 export async function newPost(req, res) {
+	// Check if the user is posting to their own author page
+	if (req.user.id !== req.params.authorId) {
+		return res.status(403).json({ error: 'Cannot post on this profile' });
+	}
+
 	const post = req.body;
+	post.authorId = req.user.id;
 	post.id = cuid();
 	post.published = new Date();
